backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
python-dotenv
django-cors-headers
gunicorn

pyicloud
google-auth-oauthlib
requests
Pillow
